LayoutTests/http/tests/security/contentSecurityPolicy/same-origin-plugin-document-blocked-in-child-window.html

<!DOCTYPE html>
<html>
<head>
<meta http-equiv="Content-Security-Policy" content="script-src 'self' 'unsafe-inline'; object-src 'none'">
<script>
if (window.testRunner) {
    testRunner.dumpAsText();
    testRunner.setCanOpenWindows();
    testRunner.setCloseRemainingWindowsWhenComplete(true);
    testRunner.waitUntilDone();
}
</script>
</head>
<body>
<script>
checkDidLoadAndNotifyDone(window.open("http://127.0.0.1:8000/plugins/resources/plugin-document-alert-and-notify-done.pl"));

// Assumes that the specified window is loading a same-origin document.
function checkDidLoadAndNotifyDone(childWindow)
{
    function checkDidLoad() {
        if (childWindow.document.location.origin !== document.location.origin)
            return;
        // Child window did load
        window.clearInterval(intervalID);
        if (window.testRunner)
            testRunner.notifyDone();
    }
    intervalID = window.setInterval(checkDidLoad, 10);
}
</script>
</body>
</html>
